move to package and rewrite a few small things
