+++ belda/beldarooms.py
@@ -1,0 +1,368 @@
+room_list = [
+
+    ######################################################### FIRST ROW #########################################################
+
+    [
+        # 0
+        "1111111111111111",
+        "1111111111111111",
+        "111111111,,,1111",
+        "111111,,,,,,,111",
+        "11111,,,11,,,,11",
+        "1,,,,,,11S,,,,,,",
+        "1,,,,,,,,,,,,,11",
+        "111,,,,1,,,,,111",
+        "1111,,11,,,11111",
+        "111,,,,,,,,,,111",
+        "1111111,,1111111"],
+    [
+        # 1
+        "1111111111111111",
+        "1111111111111111",
+        "11111111t...1111",
+        "1,,,,,,,...'..11",
+        "1,1111,;........",
+        ",,11t...11..;.'.",
+        "1111............",
+        "111..1....'..t.1",
+        "11.'.11.;.111111",
+        "1.....'.11111111",
+        "1111111111111111"],
+    [
+        # 2
+        "1111111111111111",
+        "1111111111C11111",
+        "11111111t.;.1111",
+        "1.tttt'.....'.11",
+        "...t.....'.....1",
+        ".'....tt....;..1",
+        "......1111111.'1",
+        "1.;...11.t111;;1",
+        "1...;'.1;..11;'1",
+        "1......1...'..;1",
+        "11111111..111111"],
+    [
+        # 3
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 4
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 5
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 6
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 7
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 8
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 9
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+
+        ######################################################### SECOND ROW #########################################################
+
+        # 10
+        "1111111,,1111111",
+        "1111,,,,,,,,,,,1",
+        "11,,,,11,,,,,,,1",
+        "1,,,,,11,,,,,111",
+        "1,,,,,,,,,,,,111",
+        "1,,11,,,,,,,,,,,",
+        "1,,11,,,,,,,,,,1",
+        "1,,,,,,,,111,,,1",
+        "1,,,,,,,,111,,,1",
+        "1111,,,,,,,,,,11",
+        "1111111111111111"],
+
+    [
+        # 11
+        "1111111111111111",
+        "1tt.;tt...tt..tt",
+        "1t.'....;.t...tt",
+        "1...t111..''....",
+        "1.;..1C1....'...",
+        "....11.....;...t",
+        "111111.'.....'.t",
+        "1t.'tt.....t....",
+        "1t.....;.tttt...",
+        "1ttt;..ttttt;..t",
+        "1ttttt.ttttttttt"],
+    [
+        # 12
+        "tttttttt..tttttt",
+        "tttt..;...;tt;tt",
+        "tt.;';..'....'.t",
+        "........11..;..t",
+        "..;...t111....'t",
+        "tttttttt11.;;'tt",
+        "ttttt'.tt..'..tt",
+        ".'.tt........'.t",
+        "....;....'..;;tt",
+        "t.;;...t....tttt",
+        "tttttttt..tttttt"],
+    [
+        # 13
+        "1111,,1111111111",
+        "t.,,,,,,,1111111",
+        "t..,,,,,,,,11111",
+        "t........,,11C11",
+        "t.t...;;....;..t",
+        "t.tt'tt...tt...t",
+        "t.tt.ttt..tt.'.t",
+        "t.tt.ttt'......t",
+        "t;tt..tt...'....",
+        "t.....;.....tttt",
+        "tttttttt...ttttt"],
+    [
+        # 14
+        "tttttttttttttttt",
+        "tttt........tttt",
+        "ttt...........tt",
+        "t..............t",
+        "t..............t",
+        "t......tt.......",
+        "t.....tttt......",
+        "t.......t.......",
+        "...............t",
+        "tt............tt",
+        "ttt.........tttt"],
+    [
+        # 15
+        "tttttttttttttttt",
+        "tttt........tttt",
+        "ttt...........tt",
+        "t..............t",
+        "t..............t",
+        ".......tt......t",
+        "......tttt.....t",
+        ".......t.......t",
+        "t...............",
+        "tt............tt",
+        "tttt.........ttt"],
+    [
+        # 16
+        "tttttttttt.ttttt",
+        "tttttt..t...t..t",
+        "tttttt.........t",
+        "tttttt..t...t..t",
+        "tttttt.........t",
+        "tttttt..t...t..t",
+        "tttttt.........t",
+        "tttttt..t...t..t",
+        "...............t",
+        "tttttttttttttttt",
+        "tttttttttttttttt"],
+    [
+        # 17
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 18
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+        # 19
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111111111111"],
+    [
+
+        ######################################################### THIRD ROW #########################################################
+
+        # 20
+        "1111111111111111",
+        "111,,,,,,,,,,,,,",
+        "11,,111111111111",
+        "11,,,,1111C11111",
+        "1111,,.111,,1111",
+        "111,,,,11,,11111",
+        "1111,,,,,,,111C1",
+        "11111,,,,,111,,1",
+        "1111111111111,,1",
+        "1111111111,,,,11",
+        "1111111111,11111"],
+    [
+        # 21
+        "1ttttt.ttttttttt",
+        "...;'..tttt...;.",
+        "1t..tttttt..'...",
+        "1ttttt.;;...t...",
+        "1...;.....wwwwww",
+        "1.twwwww..wwwwww",
+        "1..wwwwwwwwwwwww",
+        "1;.wwwwwwwwwwwww",
+        "1.'wwwwwwwwwwwww",
+        "1.;wwwwwwwwwwwww",
+        "1..wwwwwwwwwwwww"],
+    [
+        # 22
+        "tttttttt..tttttt",
+        "t.;..ttt...;....",
+        ".....;.......'..",
+        ".........''....t",
+        "wwwwwwwwwwwww..t",
+        "wwwwwwwwwwwww.;t",
+        "wwwwwwwwwwwww..t",
+        "wwwwwwwwwwwww..t",
+        "wwwwwwwwwwwww'.t",
+        "wwwwwwwwwwwww..t",
+        "wwwwwwwwwwwww.'t"],
+    [
+        # 23
+        "tttttttt..tttttt",
+        ".....tt.....t..t",
+        "...............t",
+        "tt.............t",
+        "tttttt.....ttttt",
+        "ttttttt...tttttt",
+        "tt..ttt..ttttttt",
+        "tt...tt.tttttttt",
+        "tt......tttttttt",
+        "ttttttt.tttttttt",
+        "ttttttt.tttttttt"],
+    [
+        # 24
+        "ttt.........tttt",
+        "tt...........ttt",
+        "t..............t",
+        "t.......t.......",
+        "t.....tttt......",
+        "t......tt.......",
+        "t..............t",
+        "t..............t",
+        "tt...........ttt",
+        "tttt........tttt",
+        "tttttttttttttttt"],
+    [
+        # 25
+        "tttt.........ttt",
+        "ttt...........tt",
+        "t....wwwwwww....",
+        "....wwwwwwwww..t",
+        "....www.c.www..t",
+        "....www...www..t",
+        "t...wwww.wwww..t",
+        "t....wwwwwww...t",
+        "ttt...........tt",
+        "tttt........tttt",
+        "tttttttttttttttt"],
+]
+
+cave_list = [
+
+    [
+        # 1
+        "1111111111111111",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1........S.....1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1..............1",
+        "1111111..1111111"],
+
+]
