+++ belda/tiles.py
@@ -5,10 +5,8 @@ from settings import TILESIZE
 
 
 class Tile(Sprite):
-    def __init__(self, game, x, y, group, image):
-        self.groups = group
-        Sprite.__init__(self, self.groups)
-        self.game = game
+    def __init__(self, x, y, group, image):
+        Sprite.__init__(self, group)
         self.image = image
         self.image = transform.scale(self.image, (TILESIZE, TILESIZE))
         self.rect = self.image.get_rect()
